requirements.txt
# --- Core ---
Django==5.2.7
gunicorn==22.0.0
whitenoise==6.6.0

# --- Database (PostgreSQL) ---
psycopg[binary]==3.1.19
dj-database-url==2.2.0

# --- Optional: API (DRF) ---
djangorestframework==3.15.2

# --- Optional: OIDC (Entra ID via mozilla-django-oidc) ---
mozilla-django-oidc==3.0.0

# --- Optional: Background tasks (Celery/Redis) ---
celery==5.3.6
redis==5.0.6
brotli==1.1.0
